+++ docs/changelog.rst
@@ -6,6 +6,7 @@ ChangeLog
 0.5
 ---
 
+- made urls Django 1.5 compatible
 - add maximum_choices field
 0.4.2
 -----